# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Berita Documentation",
        "description": "Dokumentasi API untuk sistem manajemen berita",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/reports": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Get laporan user (khusus admin)",
                "description": "Mengambil daftar laporan untuk admin review",
                "operationId": "81a8afd2aa3c69c6b5565682c3a78dd7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "reviewed",
                                "resolved"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List laporan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Report"
                ],
                "summary": "Laporkan berita",
                "description": "Melaporkan berita yang dianggap melanggar aturan",
                "operationId": "39a2e936857d7fa55026c1daac0f01d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_berita",
                                    "reason"
                                ],
                                "properties": {
                                    "id_berita": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reason": {
                                        "type": "string",
                                        "enum": [
                                            "spam",
                                            "konten_tidak_pantast",
                                            "hoax",
                                            "pelanggaran_hak_cipta",
                                            "lainnya"
                                        ],
                                        "example": "hoax"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Berita ini mengandung informasi yang salah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Laporan berhasil dikirim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/{id_report}": {
            "put": {
                "tags": [
                    "Report"
                ],
                "summary": "Update status laporan (khusus admin)",
                "description": "Admin mengupdate status laporan",
                "operationId": "65c250474ec1be507d0255180418b5ce",
                "parameters": [
                    {
                        "name": "id_report",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "reviewed",
                                            "resolved"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status laporan berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search berita berdasarkan query",
                "description": "Pencarian berita berdasarkan judul, isi, atau penulis",
                "operationId": "267256bf9c306b2e709354696ddb8941",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query pencarian",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "politik"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah hasil per halaman",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil pencarian",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_berita": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "judul": {
                                                "type": "string",
                                                "example": "Berita Politik Terbaru"
                                            },
                                            "isi": {
                                                "type": "string",
                                                "example": "Isi berita..."
                                            },
                                            "penulis": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/search/advanced": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Pencarian advanced berita",
                "description": "Pencarian berita dengan filter kategori, penulis, tanggal",
                "operationId": "5d4f10f697f8f31f8ab1622b6f9f00c0",
                "parameters": [
                    {
                        "name": "judul",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "kategori",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "penulis",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tanggal_dari",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "tanggal_sampai",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil pencarian advanced"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT"
            }
        }
    },
...
}